requirements.txt
pyautogui>=0.9.54
pyserial>=3.5
pillow>=10.0.0
